Address review: recovery, endpoint scoping, chain invalidation
- One-shot session-scoped recovery when previous_response_id is rejected,
  so a dead parent cannot wedge the session permanently.
- Parent lookup rejects a KNOWN-different endpoint but tolerates an absent
  providerBaseURL, which stampAiTurnModel legitimately omits.
- Compression and density mutations invalidate the stateful chain.
- previous_response_id is executor-owned and always cleared before reuse.
- Split oversized test file and share its harness rather than duplicate it.
