+++ packages/providers/src/openai-responses/__tests__/OpenAIResponsesProvider.codex.stateful.remediation.test.ts
@@ -16,31 +16,22 @@
  */
 
 /**
- * Issue #3134 remediation tests — behavioral tests for the 8 review fixes.
+ * Chain invalidation, request invariants, and end-to-end chaining.
  *
- * These tests exercise real code paths: the transport derives its metadata
- * from a REAL SSE byte stream through parseResponsesStream (using the
- * completingScript harness), and turn-2 history is fed verbatim from turn-1's
- * ACTUAL yielded IContents, proving the stateful chain works end-to-end.
+ * Behavioral tests for issue #3134 Codex statefulness. The transport derives
+ * its metadata from a REAL SSE byte stream through parseResponsesStream, and
+ * turn-2 history is fed verbatim from turn-1's ACTUAL yielded IContents.
  */
 
 import { afterEach, beforeEach, describe, expect, it } from 'bun:test';
-import { SettingsService } from '@vybestack/llxprt-code-settings';
 import {
   clearActiveProviderRuntimeContext,
   createProviderRuntimeContext,
   setActiveProviderRuntimeContext,
 } from '@vybestack/llxprt-code-core/runtime/providerRuntimeContext.js';
-import { createRuntimeInvocationContext } from '@vybestack/llxprt-code-core/runtime/RuntimeInvocationContext.js';
-import { createRuntimeConfigStub } from '@vybestack/llxprt-code-core/test-utils/runtime.js';
-import { createProviderCallOptions } from '@vybestack/llxprt-code-core/test-utils/providerCallOptions.js';
+import { SettingsService } from '@vybestack/llxprt-code-settings';
 import type { IContent } from '@vybestack/llxprt-code-core/services/history/IContent.js';
 import { OpenAIResponsesProvider } from '../OpenAIResponsesProvider.js';
-import type { OpenAIResponsesRequest } from '../OpenAIResponsesTypes.js';
-import type {
-  StreamResponseOptions,
-  WebSocketTransport,
-} from '../openAIResponsesWebSocketTransport.js';
 import {
   SocketHarness,
   completingScript,
@@ -48,139 +39,20 @@ import {
   userTextsOf,
 } from '../openAIResponsesWebSocketTransport.test-helpers.js';
 import { createCodexResponsesWebSocketTransport } from '../openAIResponsesWebSocketTransport.js';
+import { executeOpenAIResponsesRequest } from '../openAIResponsesExecutor.js';
+import { createRuntimeConfigStub } from '@vybestack/llxprt-code-core/test-utils/runtime.js';
+import { createRuntimeInvocationContext } from '@vybestack/llxprt-code-core/runtime/RuntimeInvocationContext.js';
+import { createProviderCallOptions } from '@vybestack/llxprt-code-core/test-utils/providerCallOptions.js';
 import {
-  executeOpenAIResponsesRequest,
-  type ResponsesExecutorDeps,
-} from '../openAIResponsesExecutor.js';
-import type { NormalizedGenerateChatOptions } from '../../BaseProvider.js';
+  CODEX_BASE_URL,
+  TEST_RUNTIME_ID,
+  buildCodexOAuthManager,
+  buildDeps,
+  buildOptions,
+  metadataOf,
+} from '../codexStateful.test-helpers.js';
 
-const CODEX_BASE_URL = 'https://chatgpt.com/backend-api/codex';
-const TEST_RUNTIME_ID = 'codex-stateful-remediation-test';
-
-function buildCodexOAuthManager(): object {
-  return {
-    getOAuthToken: async () => ({
-      access_token: 'codex-token',
-      token_type: 'Bearer',
-      expires_in: 3600,
-      expiry: Math.floor(Date.now() / 1000) + 3600,
-      refresh_token: 'test-refresh',
-      scope: 'openid',
-      account_id: 'acct_codex_123',
-    }),
-  };
-}
-
-function sseResponse(
-  id: string,
-  text: string,
-  responsesStored: boolean,
-): Response {
-  const encoder = new TextEncoder();
-  const body = new ReadableStream<Uint8Array>({
-    start(controller) {
-      if (text) {
-        controller.enqueue(
-          encoder.encode(
-            `data: {"type":"response.output_text.delta","delta":${JSON.stringify(text)}}\n\n`,
-          ),
-        );
-      }
-      const stored = responsesStored ? ',"responsesStored":true' : '';
-      void stored;
-      controller.enqueue(
-        encoder.encode(
-          `data: {"type":"response.completed","response":{"id":${JSON.stringify(id)},"status":"completed"}}\n\n`,
-        ),
-      );
-      controller.enqueue(encoder.encode('data: [DONE]\n\n'));
-      controller.close();
-    },
-  });
-  return new Response(body, {
-    status: 200,
-    headers: { 'content-type': 'text/event-stream' },
-  });
-}
-
-function errorResponse(status: number, message: string): Response {
-  const body = JSON.stringify({
-    error: { message, type: 'invalid_request_error' },
-  });
-  return new Response(body, {
-    status,
-    headers: { 'content-type': 'application/json' },
-  });
-}
-
-async function drain(
-  iterator: AsyncIterableIterator<IContent>,
-): Promise<IContent[]> {
-  const out: IContent[] = [];
-  for await (const chunk of iterator) out.push(chunk);
-  return out;
-}
-
-function buildOptions(
-  contents: IContent[],
-  ephemerals: Record<string, unknown> = {},
-): NormalizedGenerateChatOptions {
-  const settings = new SettingsService();
-  const runtime = createProviderRuntimeContext({
-    settingsService: settings,
-    runtimeId: TEST_RUNTIME_ID,
-  });
-  const invocation = createRuntimeInvocationContext({
-    runtime,
-    settings,
-    providerName: 'openai-responses',
-    ephemeralsSnapshot: ephemerals,
-  });
-  return createProviderCallOptions({
-    providerName: 'openai-responses',
-    settings,
-    config: createRuntimeConfigStub(settings),
-    runtime,
-    invocation,
-    contents,
-    ephemeralSettings: ephemerals,
-    resolved: {
-      model: 'gpt-5.6-sol',
-      baseURL: CODEX_BASE_URL,
-      authToken: 'test-token',
-    },
-  });
-}
-
-function buildDeps(
-  overrides: Partial<ResponsesExecutorDeps> = {},
-): ResponsesExecutorDeps {
-  return {
-    providerName: 'openai-responses',
-    logger: {
-      debug: () => undefined,
-    } as unknown as ResponsesExecutorDeps['logger'],
-    getProviderBaseURL: () => CODEX_BASE_URL,
-    getCustomHeaders: () => ({ 'X-Provider': 'p' }),
-    isCodexBaseURL: (url) => (url ?? '').includes('backend-api/codex'),
-    getCodexAccountId: async () => 'codex-account',
-    resolveAuthTokenForPrompt: async () => 'codex-token',
-    generateSyntheticCallId: () => 'call_synthetic_test',
-    shouldRetryOnError: () => false,
-    getDefaultModel: () => 'gpt-5.6-sol',
-    getGlobalConfig: () => undefined,
-    ...overrides,
-  };
-}
-
-function metadataOf(messages: readonly IContent[]): IContent['metadata'] {
-  const found = messages.find(
-    (m) => m.metadata !== undefined && m.blocks.length === 0,
-  );
-  return found?.metadata;
-}
-
-describe('OpenAIResponsesProvider Codex stateful remediation @issue:3134', () => {
+describe('OpenAIResponsesProvider Codex stateful — chain invalidation, request invariants and end-to-end chaining @issue:3134', () => {
   beforeEach(() => {
     setActiveProviderRuntimeContext(
       createProviderRuntimeContext({
@@ -192,247 +64,6 @@ describe('OpenAIResponsesProvider Codex stateful remediation @issue:3134', () =>
 
   afterEach(() => {
     clearActiveProviderRuntimeContext();
-  });
-
-  describe('Fix 1 — one-shot recovery when previous_response_id is rejected', () => {
-    it('retries once with full history when the API rejects a stored parent', async () => {
-      let fetchCalls = 0;
-      let firstBody: string | undefined;
-      let secondBody: string | undefined;
-      const originalFetch = globalThis.fetch;
-      (globalThis as { fetch: unknown }).fetch = async (
-        _input: unknown,
-        init?: RequestInit,
-      ) => {
-        fetchCalls += 1;
-        const blob = init?.body;
-        const bodyText =
-          blob instanceof Blob ? await blob.text() : String(blob ?? '');
-        if (fetchCalls === 1) firstBody = bodyText;
-        else secondBody = bodyText;
-        if (fetchCalls === 1) {
-          return errorResponse(
-            400,
-            "Previous response with id 'resp_dead' not found",
-          );
-        }
-        return sseResponse('resp_recovered', 'recovered text', true);
-      };
-      try {
-        const contents: IContent[] = [
-          { speaker: 'human', blocks: [{ type: 'text', text: 'first q' }] },
-          {
-            speaker: 'ai',
-            blocks: [{ type: 'text', text: 'first a' }],
-            metadata: {
-              id: 'resp_dead',
-              responsesStored: true,
-              providerBaseURL: CODEX_BASE_URL,
-            },
-          },
-          { speaker: 'human', blocks: [{ type: 'text', text: 'second q' }] },
-        ];
-
-        let statefulFailed = false;
-        const deps = buildDeps({
-          isResponsesStatefulFailed: () => statefulFailed,
-          markResponsesStatefulFailed: () => {
-            statefulFailed = true;
-          },
-        });
-        const messages = await drain(
-          executeOpenAIResponsesRequest(buildOptions(contents), deps),
-        );
-
-        expect(fetchCalls).toBe(2);
-        expect(statefulFailed).toBe(true);
-
-        const firstReq = JSON.parse(firstBody!) as Record<string, unknown>;
-        expect(firstReq['previous_response_id']).toBe('resp_dead');
-        const firstUsers = userTextsOf(firstReq['input']);
-        expect(firstUsers).not.toContain('first q');
-
-        const secondReq = JSON.parse(secondBody!) as Record<string, unknown>;
-        expect(secondReq['previous_response_id']).toBeUndefined();
-        const secondUsers = userTextsOf(secondReq['input']);
-        expect(secondUsers).toContain('first q');
-        expect(secondUsers).toContain('second q');
-
-        const text = messages
-          .flatMap((m) => m.blocks)
-          .map((b) => (b.type === 'text' ? b.text : ''))
-          .join('');
-        expect(text).toContain('recovered text');
-      } finally {
-        (globalThis as { fetch: unknown }).fetch = originalFetch;
-      }
-    });
-
-    it('after recovery, subsequent turns suppress statefulness (session-scoped flag)', async () => {
-      let fetchCalls = 0;
-      const originalFetch = globalThis.fetch;
-      (globalThis as { fetch: unknown }).fetch = async () => {
-        fetchCalls += 1;
-        if (fetchCalls === 1) {
-          return errorResponse(
-            400,
-            "previous_response_id 'resp_dead' does not exist",
-          );
-        }
-        return sseResponse('resp_new', 'ok', true);
-      };
-      try {
-        let statefulFailed = false;
-        const deps = buildDeps({
-          isResponsesStatefulFailed: () => statefulFailed,
-          markResponsesStatefulFailed: () => {
-            statefulFailed = true;
-          },
-        });
-
-        const historyWithDeadParent: IContent[] = [
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q1' }] },
-          {
-            speaker: 'ai',
-            blocks: [{ type: 'text', text: 'a1' }],
-            metadata: {
-              id: 'resp_dead',
-              responsesStored: true,
-              providerBaseURL: CODEX_BASE_URL,
-            },
-          },
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q2' }] },
-        ];
-        await drain(
-          executeOpenAIResponsesRequest(
-            buildOptions(historyWithDeadParent),
-            deps,
-          ),
-        );
-
-        // Turn 2: even though history now has resp_new with responsesStored,
-        // the session flag suppresses statefulness.
-        fetchCalls = 0;
-        let turn2Body: string | undefined;
-        (globalThis as { fetch: unknown }).fetch = async (
-          _input: unknown,
-          init?: RequestInit,
-        ) => {
-          const blob = init?.body;
-          turn2Body =
-            blob instanceof Blob ? await blob.text() : String(blob ?? '');
-          return sseResponse('resp_new2', 'ok2', true);
-        };
-
-        const historyAfterRecovery: IContent[] = [
-          ...historyWithDeadParent,
-          {
-            speaker: 'ai',
-            blocks: [{ type: 'text', text: 'a2' }],
-            metadata: {
-              id: 'resp_new',
-              responsesStored: true,
-              providerBaseURL: CODEX_BASE_URL,
-            },
-          },
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q3' }] },
-        ];
-        await drain(
-          executeOpenAIResponsesRequest(
-            buildOptions(historyAfterRecovery),
-            deps,
-          ),
-        );
-
-        const req = JSON.parse(turn2Body!) as Record<string, unknown>;
-        expect(req['previous_response_id']).toBeUndefined();
-        const users = userTextsOf(req['input']);
-        expect(users).toContain('q1');
-        expect(users).toContain('q3');
-      } finally {
-        (globalThis as { fetch: unknown }).fetch = originalFetch;
-      }
-    });
-  });
-
-  describe('Fix 2 — parent lookup must match endpoint', () => {
-    it('does not select a parent from a different providerBaseURL', async () => {
-      const harness = new SocketHarness([completingScript('ok')]);
-      const transport = createCodexResponsesWebSocketTransport({
-        openSocket: harness.openSocket,
-      });
-      try {
-        const contents: IContent[] = [
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q1' }] },
-          {
-            speaker: 'ai',
-            blocks: [{ type: 'text', text: 'a1' }],
-            metadata: {
-              id: 'resp_other',
-              responsesStored: true,
-              providerBaseURL: 'https://api.openai.com/v1',
-            },
-          },
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q2' }] },
-        ];
-        await drainHarness(
-          executeOpenAIResponsesRequest(
-            buildOptions(contents),
-            buildDeps({ getWebSocketTransport: () => transport }),
-          ),
-        );
-
-        const sent = JSON.parse(harness.sockets[0].sent[0]) as Record<
-          string,
-          unknown
-        >;
-        expect(sent['previous_response_id']).toBeUndefined();
-        const users = userTextsOf(sent['input']);
-        expect(users).toContain('q1');
-        expect(users).toContain('q2');
-      } finally {
-        transport.close();
-      }
-    });
-
-    it('selects a parent when providerBaseURL matches the resolved endpoint', async () => {
-      const harness = new SocketHarness([completingScript('ok')]);
-      const transport = createCodexResponsesWebSocketTransport({
-        openSocket: harness.openSocket,
-      });
-      try {
-        const contents: IContent[] = [
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q1' }] },
-          {
-            speaker: 'ai',
-            blocks: [{ type: 'text', text: 'a1' }],
-            metadata: {
-              id: 'resp_match',
-              responsesStored: true,
-              providerBaseURL: CODEX_BASE_URL,
-            },
-          },
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q2' }] },
-        ];
-        await drainHarness(
-          executeOpenAIResponsesRequest(
-            buildOptions(contents),
-            buildDeps({ getWebSocketTransport: () => transport }),
-          ),
-        );
-
-        const sent = JSON.parse(harness.sockets[0].sent[0]) as Record<
-          string,
-          unknown
-        >;
-        expect(sent['previous_response_id']).toBe('resp_match');
-        const users = userTextsOf(sent['input']);
-        expect(users).not.toContain('q1');
-        expect(users).toContain('q2');
-      } finally {
-        transport.close();
-      }
-    });
   });
 
   describe('Fix 3 — compression/density strips responsesStored', () => {
@@ -541,78 +172,6 @@ describe('OpenAIResponsesProvider Codex stateful remediation @issue:3134', () =>
         // No stored parent → no previous_response_id should be sent, even
         // though the user supplied one.
         expect(sent['previous_response_id']).toBeUndefined();
-      } finally {
-        transport.close();
-      }
-    });
-  });
-
-  describe('Fix 6 — AGENTS.md injection only on no-parent turns', () => {
-    it('injects the synthetic config read when there is no parent', async () => {
-      const harness = new SocketHarness([completingScript('ok')]);
-      const transport = createCodexResponsesWebSocketTransport({
-        openSocket: harness.openSocket,
-      });
-      try {
-        const contents: IContent[] = [
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q1' }] },
-        ];
-        await drainHarness(
-          executeOpenAIResponsesRequest(
-            buildOptions(contents),
-            buildDeps({ getWebSocketTransport: () => transport }),
-          ),
-        );
-
-        const sent = JSON.parse(harness.sockets[0].sent[0]) as Record<
-          string,
-          unknown
-        >;
-        const input = sent['input'] as Array<{ type?: string }>;
-        const hasReadFileCall = input.some(
-          (item) => item.type === 'function_call',
-        );
-        expect(hasReadFileCall).toBe(true);
-      } finally {
-        transport.close();
-      }
-    });
-
-    it('does NOT inject the synthetic config read when a parent is active', async () => {
-      const harness = new SocketHarness([completingScript('ok')]);
-      const transport = createCodexResponsesWebSocketTransport({
-        openSocket: harness.openSocket,
-      });
-      try {
-        const contents: IContent[] = [
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q1' }] },
-          {
-            speaker: 'ai',
-            blocks: [{ type: 'text', text: 'a1' }],
-            metadata: {
-              id: 'resp_parent',
-              responsesStored: true,
-              providerBaseURL: CODEX_BASE_URL,
-            },
-          },
-          { speaker: 'human', blocks: [{ type: 'text', text: 'q2' }] },
-        ];
-        await drainHarness(
-          executeOpenAIResponsesRequest(
-            buildOptions(contents),
-            buildDeps({ getWebSocketTransport: () => transport }),
-          ),
-        );
-
-        const sent = JSON.parse(harness.sockets[0].sent[0]) as Record<
-          string,
-          unknown
-        >;
-        const input = sent['input'] as Array<{ type?: string; name?: string }>;
-        const hasReadFileCall = input.some(
-          (item) => item.type === 'function_call' && item.name === 'read_file',
-        );
-        expect(hasReadFileCall).toBe(false);
       } finally {
         transport.close();
       }
